% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocu.R
\docType{data}
\name{airquality}
\alias{airquality}
\title{Daily air pollutant measures in Leeds}
\description{
The \code{airquality} data frame consists of measurements of pollutants in the city of Leeds. In the saying of Boldi and Davison (2007): 
 "The dataset comprise daily series of monitoring measurements of ozone levels (O3), nitrogen dioxide (NO2), nitrogen oxide (NO) and particulate matter (PM10), in the city centre of Leeds, UK, over 1994--1998. Levels of the gases are measured in parts per billion, and those of PM 10 in micrograms per cubic metre. "
}
\details{
Downloaded from \url{http://www.airquality.co.uk}. (c) Crown 2015 copyright Defra via uk-air.defra.gov.uk, licenced under the Open Government Licence (OGL).
}
\references{
Heffernan, J., Tawn, J. (2004)  A conditional approach for multivariate extreme values (with discussion).   \emph{J. R. Stat. Soc., Ser. B Stat. Methodol.} \bold{66}(3), 497--546.

Sabourin, A. , Naveau, P., Fougeres, A.-L. (2013) Bayesian Model averaging for Multivariate extremes. \emph{Extremes}, \bold{16}(3), 325--350.

 Boldi, M.O., Davison, A.C. (2007)  A mixture model for multivariate extremes. \emph{J. R. Stat. Soc., Ser. B Stat. Methodol.} \bold{69}(2), 217--229.
 
 Cooley, D., Davis, R., Naveau, P. (2010)  The pairwise beta distribution: A flexible parametric multivariate model for extremes.   \emph{J. Multivar. Anal.} \bold{101}(9), 2103--2117.
}
\keyword{datasets}
