% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ctspecdens}
\alias{.ctspecdens}
\title{Spectral density of Coles and Tawn extreme value distribution}
\usage{
.ctspecdens(param, dat, transform = TRUE)
}
\arguments{
\item{param}{vector containing the parameters \eqn{\alpha}.}

\item{dat}{matrix of pseudo-dat, of dimension \eqn{d}}

\item{transform}{logical indicating whether parameters are on the log scale. Default to \code{TRUE}}
}
\value{
the value of the log likelihood for a sample of size \code{n}
}
\description{
Spectral density of Coles and Tawn extreme value distribution
}
