% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liouville.R
\name{.liouv.iTau_s}
\alias{.liouv.iTau_s}
\title{Moment estimate for \code{theta} derived from Kendall's \eqn{\tau}{tau}}
\usage{
.liouv.iTau_s(tau_hat, family, alphavec)
}
\arguments{
\item{tau_hat}{estimated Kendall's \eqn{\tau}{tau} value from data}

\item{family}{family of the Liouville copula. Either \code{"clayton"} or \code{"gumbel"}}

\item{alphavec}{vector of Dirichlet allocations (must be a vector of integers)}
}
\value{
Value of theta
}
\description{
The moment estimates are based on inversion of the formula Kendall's \eqn{\tau}{tau} for Clayton or Gumbel Liouville copula
}
