% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liouville.R
\name{.liouv.Tau_s}
\alias{.liouv.Tau_s}
\title{Computes Kendall's \eqn{\tau}{tau} for Clayton or Gumbel Liouville copula}
\usage{
.liouv.Tau_s(theta, family, alphavec)
}
\arguments{
\item{theta}{parameter of the corresponding Archimedean copula}

\item{family}{family of the Liouville copula. Either \code{"clayton"} or \code{"gumbel"}}

\item{alphavec}{vector of Dirichlet allocations (must be a vector of integers)}
}
\value{
value of \eqn{\tau}{tau}
}
\description{
The function computes Kendall's \eqn{\tau}{tau} for the given model, given \code{alphavec}
}
