% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liouville_extreme.R
\name{.pickands.fun.uni}
\alias{.pickands.fun.uni}
\title{Pickands dependence function for the copula domain of attraction of Liouville copulas}
\usage{
.pickands.fun.uni(t, rho = 0.5, alpha = c(1, 1))
}
\arguments{
\item{t}{pseudo-angle in (0,1)}

\item{rho}{index of regular variation parameter}

\item{alpha}{vector of Dirichlet allocations. Currently must be of length 2}
}
\value{
value of Pickands function
}
\description{
Pickands dependence function as in \cite{Belzile (2014), Proposition 41}
Returns the Pickands dependence function of the copula domain of attraction
of the copula. This is only derived and implemented in the bivariate case.
}
