% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liouville.R
\name{express_coef_clayton}
\alias{express_coef_clayton}
\title{Clayton coefficient for Kendall's tau formula}
\usage{
express_coef_clayton(i, j, theta, a)
}
\arguments{
\item{i}{first index}

\item{j}{second index}

\item{theta}{parameter of the corresponding Archimedean copula}

\item{a}{\eqn{L_1}{L1}-norm of Dirichlet allocation vector}
}
\value{
Clayton coefficient
}
\description{
The function is used internally in \cite{liouv.Tau}.
}
\keyword{internal}
