% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liouville.R
\name{express_coef_gumb}
\alias{express_coef_gumb}
\title{Gumbel coefficient for Kendall's tau formula}
\usage{
express_coef_gumb(index_k, index_l, theta, a)
}
\arguments{
\item{index_k}{first index}

\item{index_l}{second index}

\item{theta}{parameter of the corresponding Archimedean copula}

\item{a}{sum of alphavec}
}
\value{
Gumbel coefficient
}
\description{
The function is used internally in \cite{liouv.Tau}.
}
\keyword{internal}
