% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negdirspecdens}
\alias{negdirspecdens}
\title{Spectral density of scaled negative Dirichlet extreme value distribution}
\usage{
negdirspecdens(param, dat, d, transform = TRUE)
}
\arguments{
\item{param}{vector containing the parameters \eqn{\alpha} and \eqn{\rho}, in this order.}

\item{dat}{matrix of Frechet or Pareto observations, of dimension \eqn{n} by \eqn{d}.}

\item{d}{dimension of model}

\item{transform}{logical indicating whether parameters are on the log scale. Default to \code{TRUE}}
}
\value{
the log-likelihood for the \code{n} sample
}
\description{
Spectral density of scaled negative Dirichlet extreme value distribution
}
\details{
The function is provided as a wrapper and takes parameters on the log scale for \eqn{\alpha} and \eqn{\rho}.
}
