% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocu.R
\docType{data}
\name{nutrient}
\alias{nutrient}
\title{Women daily nutrient intake}
\format{This data frame contains the following columns:
\describe{
 \item{id}{A numeric vector containing the identification of the participant.}
 \item{calcium}{A numeric vector containing measured calcium.}
 \item{iron}{A numeric vector containing iron measurements.}
 \item{protein}{A numeric vector containing protein measurements.}
 \item{vitamin.a}{A numeric vector containing vitamin A measurements.}
 \item{vitamin.c}{A numeric vector containing vitamin C measurements.}
   }}
\source{
The survey data was processed by Dr. Andrew Wiesner for a course on multivariate statistics at The Pennsylvania State University.
}
\description{
The \code{nutrient} data frame consists of quintuples consisting of four day measurements for intake of calcium, iron, protein, vitamin A and C from women aged 25 to 50 in the United States as part of the "Continuing Survey of Food Intakes of Individuals" program. The processed data has
 737 measurements from a cohort study of the United States Department of Agriculture (USDA) and it is available online from the University of Pennsylvania repository.
}
\references{
Genest, C., Neslehova, J., Quessy, J.-F. (2012)  Tests of symmetry for bivariate copulas.   \emph{Annals of the Institute of Statistical Mathematics}, \bold{64}(4), 811--834.

 Genest, C. and Neslehova, J. G. (2013)  Assessing and Modeling Asymmetry in Bivariate Continuous Data.   In P. Jaworski, F. Durante, & W. K. Hardle (Eds.),   Copulae in Mathematical and Quantitative Finance, Lecture Notes in Statistics, 91--114, Springer: Berlin Heidelberg. 

Li, B., Genton, M. G. (2013) Nonparametric identification of copula structures.  \emph{Journal of the American Statistical Association}, \bold{108}(502), 666--675.
}
\keyword{datasets}
